Add files via upload
The Highest Numbered Files are the latest in progress. Some of the files are used for testing, Mem-NL.txt for 3-9 of the Numblocks is a working file. Numblocks 1-2 were scrapped, their file is Memory-NL.txt. Numblocks 6+ (I think it is 6+, maybe 7+), can run without Mem-NL, they will build it by themselves. Most of my notes are included, all in various stages of progress.